[feat] some optimizations pt.1
