+++ website/src/app/globals.css
@@ -42,17 +42,17 @@
   --radius-lg: var(--radius);
   --radius-xl: calc(var(--radius) + 4px);
 
-  --animate-aurora: aurora 60s linear infinite;
+  --animate-aurora: aurora 45s linear infinite;
   @keyframes aurora {
     from {
       background-position:
-        50% 50%,
-        50% 50%;
+        30% 45%,
+        60% 50%;
     }
     to {
       background-position:
-        350% 50%,
-        350% 50%;
+        300% 70%,
+        325% 50%;
     }
   }
 }
