+++ website/src/components/ui/aurora-background.tsx
@@ -33,11 +33,11 @@ export const AuroraBackground = ({
               "--white-gradient":
                 "repeating-linear-gradient(100deg,#fff_0%,#fff_7%,transparent_10%,transparent_12%,#fff_16%)",
 
-              "--blue-300": "#93c5fd",
-              "--blue-400": "#60a5fa",
-              "--blue-500": "#3b82f6",
-              "--indigo-300": "#a5b4fc",
-              "--violet-200": "#ddd6fe",
+              "--blue-300": "#fdcf93",
+              "--blue-400": "#fae360",
+              "--blue-500": "#8ff63b",
+              "--indigo-300": "#e2fca5",
+              "--violet-200": "#dffed6",
               "--black": "#000",
               "--white": "#fff",
               "--transparent": "transparent",
